Driver/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==25.1.0
audioread==3.0.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
filelock==3.17.0
Flask==3.1.0
flatbuffers==20181003210633
fonttools==4.56.0
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
humanfriendly==10.0
idna==3.10
imutils==0.5.4
itsdangerous==2.2.0
jax==0.5.2
jaxlib==0.5.1
Jinja2==3.1.5
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.18
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
netifaces==0.10.6
networkx==3.4.2
noisereduce==3.0.3
numba==0.61.0
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.0
onnxslim==0.1.48
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
srt==3.5.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tflite-runtime==2.14.0
threadpoolctl==3.6.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.83
ultralytics-thop==2.0.14
urllib3==2.3.0
vosk==0.3.45
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
